refactor: consolidate client functions and event listeners into respective files
