+++ public/scripts/client-helpers.js
@@ -1,3 +1,5 @@
+// INPUT: string
+// OUTPUT: string
 export function formatNameForDisplay(name) {
   if (!name) return '';
   return name.charAt(0).toUpperCase() + name.slice(1);
