+++ views/partials/footer.ejs
@@ -10,7 +10,7 @@
     const currentUserID = "<%= locals.user? user.id : null %>";
   </script>
   <script type="module" src="/scripts/functions.js"></script>
-  <script type="module" src="/scripts/buttons.js"></script>
+  <script type="module" src="/scripts/event-listeners.js"></script>
   <script type="module" src="/scripts/book-api.js"></script>
 </body>
