+++ views/partials/searchbar.ejs
@@ -21,7 +21,7 @@
     </div>
   </div>
   <div class="searchbar-button">
-    <button onclick="searchFunction(event)" data-books='<%= JSON.stringify(locals.data) %>' id="search-button" class="main-button btn-sm searchbar-button minimize-txt" type="submit">
+    <button type="button" data-books='<%= JSON.stringify(locals.data) %>' id="search-button" class="main-button btn-sm searchbar-button minimize-txt">
       Apply
     </button>
   </div>
